% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_zzz.R
\name{bs_or_bar}
\alias{bs_or_bar}
\title{beeswarm or bar plot to show the intensity, used by shiny app}
\usage{
bs_or_bar(x, col, f, diffPheno)
}
\arguments{
\item{x}{numerical value}

\item{col}{color}

\item{f}{value group}

\item{diffPheno}{pheno to distinguish}
}
\description{
beeswarm or bar plot to show the intensity, used by shiny app
}
