% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plotEIC.R
\name{colCode}
\alias{colCode}
\title{prepare input for plotEIC function}
\usage{
colCode(x, diffPheno = NULL, select = NULL)
}
\arguments{
\item{x}{a simple intensity table representing chromatogram, it should be 
a data.frame with three columns named "mz", "intensity" and file (integer)}

\item{diffPheno}{a vector to indicate which file belongs to which phenotype}

\item{select}{integer; which file is selected, will be shown in red}
}
\description{
prepare input for plotEIC function
}
\examples{
x <- readRDS(system.file(package="xcmsViewerApp", "extdata", "00_exampleData_processed.RDS"))
mtable <- x$scanMetaTab
itable <- x$scanIntensityTab
chrom <- eic(itab=itable, mtab=mtable, rt = c(0, Inf), mz=c(225, 255), na.rm = TRUE)

cc1 <- colCode(chrom, diffPheno=c(1:2), select = NULL) 
plotEIC(chrom, unifile=cc1$unifile, col=cc1$col, diffPheno=cc1$diffPheno, vline=NULL)

cc2 <- colCode(chrom, diffPheno=c(1:2), select = 1)   
plotEIC(chrom, unifile=cc2$unifile, col=cc2$col, diffPheno=cc2$diffPheno, vline=NULL)
}
