% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_eic.R
\name{eic}
\alias{eic}
\title{function to extract ms1 XICs given a simple intensity table format}
\usage{
eic(itab, mtab, rt, mz, file, na.rm = TRUE)
}
\arguments{
\item{itab}{an intensity table}

\item{mtab}{the spectra meta table of the experiment}

\item{rt}{range of retention time}

\item{mz}{range of mz}

\item{file}{which file to extract}

\item{na.rm}{logical, whether remove NA values}
}
\description{
function to extract ms1 XICs given a simple intensity table format
}
\examples{
x <- readRDS(system.file(package="xcmsViewerApp", "extdata", "00_exampleData_processed.RDS"))
mtable <- x$scanMetaTab
itable <- x$scanIntensityTab
chrom <- eic(itab=itable, mtab=mtable, rt = c(0, Inf), mz=c(225, 255), file=1, na.rm = TRUE)
}
