% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_zzz.R
\name{parseRange}
\alias{parseRange}
\title{parse rt and mz range used by shiny app}
\usage{
parseRange(x, expand = 0.05)
}
\arguments{
\item{x}{a character to be parsed, such as "123-133" or "133"}

\item{expand}{when x is parsed to a single value, how to expand the range}
}
\description{
parse rt and mz range used by shiny app
}
